+++ packages/ckbtc/src/mocks/minter.mock.ts
@@ -4,5 +4,9 @@ export const minterCanisterIdMock: Principal = Principal.fromText(
   "q3fc5-haaaa-aaaaa-aaahq-cai",
 );
 
+export const bitcoinCanisterIdMock: Principal = Principal.fromText(
+  "ghsi2-tqaaa-aaaan-aaaca-cai",
+);
+
 export const bitcoinAddressMock =
   "bcrt1qu2aqme90t6hpac50x0xw8ljwqs250vn6tzlmsv";